feat: introduce partial classes for rabbitmq
